--- carbon-aware-starter/src/main/java/com/carbonaware/apis/CarbonEmissionsParams.java
@@ -1,6 +1,0 @@
-package com.carbonaware.apis;
-
-public interface CarbonEmissionsParams {
-
-    String getLocation();
-}
