--- carbon-aware-starter/src/main/java/com/carbonaware/sci/ConfiguredEmbodiedEmissionsProvider.java
@@ -1,15 +1,0 @@
-package com.carbonaware.sci;
-
-public class ConfiguredEmbodiedEmissionsProvider implements EmbodiedEmissionsProvider {
-
-    private double embodiedEmissions;
-
-    public ConfiguredEmbodiedEmissionsProvider(double embodiedEmissions) {
-        this.embodiedEmissions = embodiedEmissions;
-    }
-
-    @Override
-    public double getEmbodiedEmissions() {
-        return embodiedEmissions;
-    }
-}
