--- carbon-aware-starter/src/main/java/com/carbonaware/sci/EmbodiedEmissionsProvider.java
@@ -1,7 +1,0 @@
-package com.carbonaware.sci;
-
-public interface EmbodiedEmissionsProvider {
-
-    double getEmbodiedEmissions();
-
-}
