--- carbon-aware-starter/src/main/java/com/carbonaware/sci/EnergyConsumptionProvider.java
@@ -1,7 +1,0 @@
-package com.carbonaware.sci;
-
-public interface EnergyConsumptionProvider {
-
-    double getEnergyConsumption();
-
-}
