--- carbon-aware-starter/src/main/java/com/carbonaware/sci/SoftwareCarbonIntensity.java
@@ -1,31 +1,0 @@
-package com.carbonaware.sci;
-
-public class SoftwareCarbonIntensity {
-
-    private double energyConsumption;
-    private double marginalEmissions;
-    private double embodiedEmissions;
-
-    public SoftwareCarbonIntensity(double energyConsumption, double marginalEmissions, double embodiedEmissions) {
-        this.energyConsumption = energyConsumption;
-        this.marginalEmissions = marginalEmissions;
-        this.embodiedEmissions = embodiedEmissions;
-    }
-
-    public double getEnergyConsumption() {
-        return energyConsumption;
-    }
-
-    public double getMarginalEmissions() {
-        return marginalEmissions;
-    }
-
-    public double getEmbodiedEmissions() {
-        return embodiedEmissions;
-    }
-
-    public double getSoftwareCarbonIntensity() {
-        return (energyConsumption * marginalEmissions) + embodiedEmissions;
-    }
-
-}
