+++ carbon-aware-starter/src/main/java/com/snowfort/carbon/apis/CarbonEmissionsParams.java
@@ -1,0 +1,6 @@
+package com.snowfort.carbon.apis;
+
+public interface CarbonEmissionsParams {
+
+    String getLocation();
+}
