+++ carbon-aware-starter/src/main/java/com/snowfort/carbon/models/Emissions.java
@@ -1,0 +1,43 @@
+package com.snowfort.carbon.models;
+
+import java.time.ZonedDateTime;
+
+public class Emissions {
+
+    private String location;
+    private ZonedDateTime time;
+    private Double rating;
+    private String duration;
+
+    public String getLocation() {
+        return location;
+    }
+
+    public void setLocation(String location) {
+        this.location = location;
+    }
+
+    public ZonedDateTime getTime() {
+        return time;
+    }
+
+    public void setTime(ZonedDateTime time) {
+        this.time = time;
+    }
+
+    public Double getRating() {
+        return rating;
+    }
+
+    public void setRating(Double rating) {
+        this.rating = rating;
+    }
+
+    public String getDuration() {
+        return duration;
+    }
+
+    public void setDuration(String duration) {
+        this.duration = duration;
+    }
+}
