+++ carbon-aware-starter/src/main/java/com/snowfort/carbon/sci/ConfiguredEmbodiedEmissionsProvider.java
@@ -1,0 +1,15 @@
+package com.snowfort.carbon.sci;
+
+public class ConfiguredEmbodiedEmissionsProvider implements EmbodiedEmissionsProvider {
+
+    private double embodiedEmissions;
+
+    public ConfiguredEmbodiedEmissionsProvider(double embodiedEmissions) {
+        this.embodiedEmissions = embodiedEmissions;
+    }
+
+    @Override
+    public double getEmbodiedEmissions() {
+        return embodiedEmissions;
+    }
+}
