+++ carbon-aware-starter/src/main/java/com/snowfort/carbon/sci/EmbodiedEmissionsProvider.java
@@ -1,0 +1,7 @@
+package com.snowfort.carbon.sci;
+
+public interface EmbodiedEmissionsProvider {
+
+    double getEmbodiedEmissions();
+
+}
