+++ carbon-aware-starter/src/main/java/com/snowfort/carbon/sci/EnergyConsumptionProvider.java
@@ -1,0 +1,7 @@
+package com.snowfort.carbon.sci;
+
+public interface EnergyConsumptionProvider {
+
+    double getEnergyConsumption();
+
+}
