+++ carbon-aware-starter/src/main/java/com/snowfort/carbon/sci/SoftwareCarbonIntensity.java
@@ -1,0 +1,31 @@
+package com.snowfort.carbon.sci;
+
+public class SoftwareCarbonIntensity {
+
+    private double energyConsumption;
+    private double marginalEmissions;
+    private double embodiedEmissions;
+
+    public SoftwareCarbonIntensity(double energyConsumption, double marginalEmissions, double embodiedEmissions) {
+        this.energyConsumption = energyConsumption;
+        this.marginalEmissions = marginalEmissions;
+        this.embodiedEmissions = embodiedEmissions;
+    }
+
+    public double getEnergyConsumption() {
+        return energyConsumption;
+    }
+
+    public double getMarginalEmissions() {
+        return marginalEmissions;
+    }
+
+    public double getEmbodiedEmissions() {
+        return embodiedEmissions;
+    }
+
+    public double getSoftwareCarbonIntensity() {
+        return (energyConsumption * marginalEmissions) + embodiedEmissions;
+    }
+
+}
