+++ carbon-aware-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.carbonaware.config.CarbonAwareAutoConfiguration
+com.snowfort.carbon.config.CarbonAwareAutoConfiguration
