--- carbon-aware-starter/src/test/java/com/carbonaware/sci/SoftwareCarbonIntensityServiceTest.java
@@ -1,6 +1,0 @@
-package com.carbonaware.sci;
-
-
-class SoftwareCarbonIntensityServiceTest {
-
-}
