+++ carbon-aware-starter/src/test/java/com/snowfort/carbon/sci/SoftwareCarbonIntensityServiceTest.java
@@ -1,0 +1,6 @@
+package com.snowfort.carbon.sci;
+
+
+class SoftwareCarbonIntensityServiceTest {
+
+}
