+++ examples/multiple-services/weather-service/pom.xml
@@ -51,7 +51,7 @@
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
 		<dependency>
-			<groupId>com.snowfort</groupId>
+			<groupId>com.snowfort.carbon</groupId>
 			<artifactId>carbon-aware-starter</artifactId>
 			<version>0.0.1-SNAPSHOT</version>
 		</dependency>
